Modulo10 - Unificação do elemento datepicker
