--- packages/GoBarber/mobile/src/components/DateInput/index.ios.js
@@ -1,41 +1,0 @@
-import React, { useState, useMemo } from 'react';
-import { DatePickerIOS } from 'react-native';
-import { format } from 'date-fns';
-import pt from 'date-fns/locale/pt';
-import Icon from 'react-native-vector-icons';
-
-import { Contianer, DateButton, DateText, Picker } from './styles';
-
-export default function DateInput({ date, onChange }) {
-  const [opened, setOpened] = useState(false);
-
-  const dateFormatted = useMemo(
-    () =>
-      format(date, "dd 'de' MMMM 'de' yyyy", {
-        locale: pt,
-      }),
-    [date]
-  );
-
-  return (
-    <Contianer>
-      <DateButton onPress={() => setOpened(!opened)}>
-        <Icon name="event" color="#fff" sixe={20} />
-        <DateText>{dateFormatted}</DateText>
-      </DateButton>
-
-      {opened && (
-        <Picker>
-          <DatePickerIOS
-            date={date}
-            onDateChange={onChange}
-            minimumDate={new Date()}
-            minuteInterval={60}
-            locale="pt"
-            mode="date"
-          />
-        </Picker>
-      )}
-    </Contianer>
-  );
-}
